requirements.txt
# Dependências principais para o backend
fastapi>=0.95.0,<0.104.0
uvicorn>=0.20.0,<0.24.0
pydantic>=1.10.0,<2.4.0  # Versão 1.x é mais estável e não exige Rust
starlette>=0.25.0,<0.28.0

# Processamento de requisições
python-multipart>=0.0.5
httpx>=0.23.0

# Banco de dados
fdb>=2.0.1  # Driver Firebird

# Autenticação e segurança
python-jose[cryptography]>=3.3.0  # Usando criptografia sem Rust
passlib[bcrypt]>=1.7.4  # Inclui bcrypt de forma pré-compilada
python-dotenv>=0.21.0

# Manipulação de datas e tempos
pytz>=2022.7

# Utilitários para manipulação de JSON e conversão de dados
json5>=0.9.10
webcolors>=1.12

# [Opcional] Ferramentas de desenvolvimento - comentadas para evitar Rust
# black>=22.12.0
# ruff>=0.0.257
